pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python_fastapi_stack"
version = "0.0.1"
description = "Python FastAPI Stack: poetry, fastapi, sqlmodel, alembic, loguru, crud, notify."
readme = "README.md"
authors = ["martokk <25113632+martokk@users.noreply.github.com>"]
license = "MIT"
repository = "https://github.com/martokk/python_fastapi_stack"
homepage = "https://github.com/martokk/python_fastapi_stack"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = ["fastapi", "poetry", "python", "sqlmodel", "alembic", "loguru", "crud"]

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = []

[tool.poetry.scripts]
# Entry points for the package https://python-poetry.org/docs/pyproject/#scripts
"python_fastapi_stack" = "python_fastapi_stack.__main__:app"

[tool.poetry.dependencies]
python = "^3.10"
loguru = "^0.6.0"
pyyaml = "^6.0"
python-dotenv = "^0.21.0"
shortuuid = "^1.0.11"
httpx = "^0.23.3"
wheel = "^0.38.1"
python-telegram-bot = "^20.0"
typer = {extras = ["all"], version = "^0.6.1"}
rich = "^12.6.0"
colorama = "^0.4.5"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"
fastapi-utils = "^0.2.1"
jinja2 = "^3.1.2"
pyjwt = "^2.6.0"
passlib = "^1.7.4"
alembic = "^1.9.2"
asyncpg = "^0.27.0"
emails = "^0.6"
python-multipart = "^0.0.5"
email-validator = "^1.3.0"
pydantic = "^1.10.4"
sqlmodel = "^0.0.8"


[tool.poetry.group.dev.dependencies]
bandit = "^1.7.1"
black = {version = "^22.3.0", allow-prereleases = true}
darglint = "^1.8.1"
flake8 = "^5.0.4"
isort = {extras = ["colors"], version = "^5.10.1"}
mypy = "^0.982"
mypy-extensions = "^0.4.3"
pre-commit = "^2.15.0"
pydocstyle = "^6.1.1"
pylint = "^2.11.1"
pytest = "^7.1.3"
pytest-mock = "^3.10.0"
pyupgrade = "^3.1.0"
safety = "^2.3.1"
coverage = "^6.1.2"
coverage-badge = "^1.1.0"
pytest-html = "^3.1.1"
pytest-cov = "^4.0.0"
pytest-asyncio = "^0.20.3"

[tool.black]
# https://github.com/psf/black
target-version = ["py310"]
line-length = 100
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | env
    | venv
    | typings
)/
'''

[tool.darglint]
# https://github.com/terrencepreilly/darglint
strictness = "long"
docstring_style = "google"
message_template="{path}  :  {line} -  {msg_id} -{msg}"


[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 310
line_length = 100

known_typing = ["typing", "typings", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true
skip = ["pyqt5_ui.py", "resources_rc.py"]
skip_glob = [".venv/*", "typings/*", "migrations/*"]

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = "3.10"
pretty = true
show_traceback = true
color_output = true
allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = false
exclude = "typings"

[[tool.mypy.overrides]]
module = "*.pyqt5_ui"
ignore_errors = true

[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
asyncio_mode="auto"

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

[tool.coverage.run]
source = ["tests"]

[coverage.paths]
source = "python_fastapi_stack"

[coverage.run]
branch = true

[coverage.report]
fail_under = 50
show_missing = true

[tool.pylint.'MESSAGES CONTROL']
extension-pkg-whitelist = "pydantic"